refactor: int 범위 이슈가 있엇고 추후 1등 상금이 int 형을 넘어 갈 수 도 있기도 하여 long으로 변경

+++ src/main/java/lotto/enums/Rank.java
@@ -12,9 +12,9 @@ public enum Rank {
 
     private final int matchCount;
     private final boolean containBonus;
-    private final int winningAmount;
+    private final long winningAmount;
 
-    Rank(int matchCount, boolean containBonus, int winningAmount) {
+    Rank(int matchCount, boolean containBonus, long winningAmount) {
         this.matchCount = matchCount;
         this.containBonus = containBonus;
         this.winningAmount = winningAmount;
@@ -28,7 +28,7 @@ public enum Rank {
         return this.containBonus;
     }
 
-    public int getWinningAmount() {
+    public long getWinningAmount() {
         return this.winningAmount;
     }
